Moves modelLineProjectionData into its own file.
