+++ src/vs/editor/common/viewModel/modelLineProjection.ts
@@ -8,7 +8,8 @@ import { Position } from 'vs/editor/common/core/position';
 import { IRange } from 'vs/editor/common/core/range';
 import { EndOfLinePreference, ITextModel, PositionAffinity } from 'vs/editor/common/model';
 import { LineInjectedText } from 'vs/editor/common/model/textModelEvents';
-import { InjectedText, ModelLineProjectionData, SingleLineInlineDecoration, ViewLineData } from 'vs/editor/common/viewModel/viewModel';
+import { InjectedText, ModelLineProjectionData } from 'vs/editor/common/viewModel/modelLineProjectionData';
+import { SingleLineInlineDecoration, ViewLineData } from 'vs/editor/common/viewModel/viewModel';
 
 export interface IModelLineProjection {
 	isVisible(): boolean;
