+++ src/vs/editor/test/common/viewModel/lineBreakData.test.ts
@@ -6,7 +6,7 @@
 import assert = require('assert');
 import { PositionAffinity } from 'vs/editor/common/model';
 import { ModelDecorationInjectedTextOptions } from 'vs/editor/common/model/textModel';
-import { ModelLineProjectionData } from 'vs/editor/common/viewModel/viewModel';
+import { ModelLineProjectionData } from 'vs/editor/common/viewModel/modelLineProjectionData';
 
 suite('Editor ViewModel - LineBreakData', () => {
 	test('Basic', () => {
